Move `@plugin` and `@config` walk into compat layer'
